+++ TransitionPaths/FHorz/subcodes/ValueFnSingleStep/QuasiHyperbolic/fastOLG/DivideConquer/GridInterpLayer/ValueFnIter_FHorz_TPath_SingleStep_QHS_fastOLG_DC1_GI1_nod_e_raw.m
@@ -118,7 +118,7 @@ if vfoptions.lowmemory==0
     Policy(3,:,:,:,:) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
     % Vunderbar = Vhat + (beta - beta0beta)*EV_at_policy
-    linidx=double(reshape(maxindexL2,[1,N_a*N_j*N_z*N_e]))+n2long*(0:N_a*N_j*N_z*N_e-1);
+    linidx=reshape(maxindexL2,[1,N_a*N_j*N_z*N_e])+n2long*(0:N_a*N_j*N_z*N_e-1);
     EV_at_policy=reshape(EVfine(linidx),[N_a*N_j,N_z,N_e]);
     Vhat=reshape(Vhat,[N_a*N_j,N_z,N_e]);
     V=Vhat+EV_at_policy;
@@ -173,7 +173,7 @@ elseif vfoptions.lowmemory==1
         inUpperStrict = (maxindexL2 >= n2short+3) & (maxindexL2 <= n2long-1);
         Policy(3,:,:,:,e_c) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-        linidx_e=double(reshape(maxindexL2,[1,N_a*N_j*N_z]))+n2long*(0:N_a*N_j*N_z-1);
+        linidx_e=reshape(maxindexL2,[1,N_a*N_j*N_z])+n2long*(0:N_a*N_j*N_z-1);
         EV_at_policy_e=reshape(EVfine_e(linidx_e),[N_a,N_j,N_z]);
         Vhat(:,:,e_c)=reshape(Vhat_e,[N_a*N_j,N_z]);
         V(:,:,:,e_c)=shiftdim(Vhat_e,1)+EV_at_policy_e;
@@ -236,7 +236,7 @@ elseif vfoptions.lowmemory==2
             inUpperStrict = (maxindexL2 >= n2short+3) & (maxindexL2 <= n2long-1);
             Policy(3,:,:,z_c,e_c) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-            linidx_ze=double(reshape(maxindexL2,[1,N_a*N_j]))+n2long*(0:N_a*N_j-1);
+            linidx_ze=reshape(maxindexL2,[1,N_a*N_j])+n2long*(0:N_a*N_j-1);
             EV_at_policy_ze=reshape(EVfine_ze(linidx_ze),[N_a,N_j]);
             Vhat(:,z_c,e_c)=reshape(Vhat_ze,[N_a*N_j,1]);
             V(:,:,z_c,e_c)=shiftdim(Vhat_ze,1)+EV_at_policy_ze;
