+++ TransitionPaths/FHorz/subcodes/ValueFnSingleStep/QuasiHyperbolic/slowOLG/DivideConquer/GridInterpLayer/ValueFnIter_FHorz_TPath_SingleStep_QHS_DC1_GI1_nod_e_raw.m
@@ -216,7 +216,7 @@ for reverse_j=1:N_j-1
         inUpperStrict = (maxindexL2 >= n2short+3) & (maxindexL2 <= n2long-1);
         Policy(3,:,:,:,jj) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-        linidx=double(reshape(maxindexL2,[1,N_a*N_z*N_e]))+n2long*(0:N_a*N_z*N_e-1);
+        linidx=reshape(maxindexL2,[1,N_a*N_z*N_e])+n2long*(0:N_a*N_z*N_e-1);
         EV_at_policy=reshape(EVfine(linidx),[N_a,N_z,N_e]);
         V(:,:,:,jj)=Vhat_jj+(beta-beta0beta)*EV_at_policy;
 
@@ -262,7 +262,7 @@ for reverse_j=1:N_j-1
             inUpperStrict = (maxindexL2 >= n2short+3) & (maxindexL2 <= n2long-1);
             Policy(3,:,:,e_c,jj) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-            linidx=double(reshape(maxindexL2,[1,N_a*N_z]))+n2long*(0:N_a*N_z-1);
+            linidx=reshape(maxindexL2,[1,N_a*N_z])+n2long*(0:N_a*N_z-1);
             EV_at_policy=reshape(EVfine(linidx),[N_a,N_z]);
             V(:,:,e_c,jj)=Vhat_jj_e+(beta-beta0beta)*EV_at_policy;
         end
@@ -310,7 +310,7 @@ for reverse_j=1:N_j-1
                 inUpperStrict = (maxindexL2 >= n2short+3) & (maxindexL2 <= n2long-1);
                 Policy(3,:,z_c,e_c,jj) = shiftdim(2 + (inLowerStrict & isInfLower) - (inUpperStrict & isInfUpper),-1);
 
-                linidx_ze=double(reshape(maxindexL2,[1,N_a]))+n2long*(0:N_a-1);
+                linidx_ze=reshape(maxindexL2,[1,N_a])+n2long*(0:N_a-1);
                 EV_at_policy_ze=reshape(EVfine_ze(linidx_ze),[N_a,1]);
                 V(:,z_c,e_c,jj)=Vhat_jj_ze+(beta-beta0beta)*EV_at_policy_ze;
             end
